requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
python-memcached==1.53
